--- user-service/src/main/kotlin/org/radarbase/user/model/User.kt
@@ -1,10 +1,0 @@
-package org.radarbase.user.model
-
-data class User(
-    val id: String,
-    val projectId: String,
-    val name: String,
-    val email: String,
-    val status: String,
-    val metadata: Map<String, Any> = emptyMap(),
-)
